refac: move methods to respective files
